Combine spark-worker-rs and spark-rs
