+++ package.json
@@ -24,7 +24,7 @@
     "build:wasm": "node rust/build_wasm.js",
     "build:watch": "onchange 'src/**/*.{ts,glsl}' -- npm run build:dev",
     "build-lod": "cargo run --manifest-path rust/build-lod/Cargo.toml --release --",
-    "clean": "rm -rf dist/ && rm -rf node_modules/ && rm -rf rust/target/ && rm -rf rust/spark-worker-rs/pkg/ && rm -rf rust/spark-rs/pkg/ && npm run assets:clean",
+    "clean": "rm -rf dist/ && rm -rf node_modules/ && rm -rf rust/target/ && rm -rf rust/spark-rs/pkg/ && npm run assets:clean",
     "dev": "npm run build:dev && (vite --host & npm run build:watch)",
     "docs": "node scripts/clean-site-files.js && node scripts/copy-site-files.js && mkdocs serve",
     "format": "biome format .",
@@ -50,15 +50,15 @@
     "lefthook": "1.11.12",
     "lil-gui": "^0.20.0",
     "onchange": "7.1.0",
+    "spark-rs": "file:rust/spark-rs/pkg",
     "stats.js": "^0.17.0",
     "three": "^0.180.0",
     "ts-node": "10.9.2",
     "typescript": "^5.7.3",
     "vite": "^6.0.11",
+    "vite-plugin-arraybuffer": "^0.1.4",
     "vite-plugin-dts": "^4.5.4",
-    "vite-plugin-glsl": "^1.3.1",
-    "spark-worker-rs": "file:rust/spark-worker-rs/pkg",
-    "spark-rs": "file:rust/spark-rs/pkg"
+    "vite-plugin-glsl": "^1.3.1"
   },
   "dependencies": {
     "fflate": "^0.8.2"
