+++ rust/.gitignore
@@ -1,3 +1,2 @@
 target/
-spark-worker-rs/pkg/
 spark-rs/pkg/
