+++ rust/spark-rs/src/decoder.rs
