+++ rust/spark-rs/src/ext_splats.rs
