--- rust/spark-worker-rs/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "spark-worker-rs"
-version = "0.1.0"
-rust-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-ahash.workspace = true
-anyhow.workspace = true
-glam.workspace = true
-half.workspace = true
-js-sys.workspace = true
-ordered-float.workspace = true
-smallvec.workspace = true
-wasm-bindgen.workspace = true
-web-sys = { workspace = true, features = ["Window", "Performance"] }
-spark-lib = { path = "../spark-lib" }
-serde-wasm-bindgen.workspace = true
-serde_json.workspace = true
-itertools.workspace = true
-console_error_panic_hook = "0.1"
